# v7: implement snapshots