This adds the concept of a `Snapshot`. This is an immutable
snapshot of the current configuration state and a
current User value that is suitable for adding to a Context
or other request-specific data.

It also adds the concept of flag types - a flag that
knows its type and can be used to extract a value
from a snapshot - usually such a flag would be
stored in a global variable. This means that
flag lookup can be a little faster because we can
avoid the map lookup.

Benchmark results compared to previous v7 code:

```
name                                   old time/op    new time/op    delta
Get/one-of/get-and-make-8                 165ns ± 1%     250ns ± 0%   +51.73%  (p=0.008 n=5+5)
Get/one-of/get-only-8                     114ns ± 4%      62ns ± 4%   -45.16%  (p=0.008 n=5+5)
Get/less-than-with-int/get-and-make-8    90.6ns ± 2%   190.9ns ± 3%  +110.72%  (p=0.008 n=5+5)
Get/less-than-with-int/get-only-8        77.0ns ± 2%    25.6ns ± 8%   -66.72%  (p=0.008 n=5+5)
Get/with-percentage/get-and-make-8        460ns ± 1%     578ns ± 9%   +25.76%  (p=0.008 n=5+5)
Get/with-percentage/get-only-8            402ns ± 2%     340ns ± 1%   -15.55%  (p=0.008 n=5+5)

name                                   old alloc/op   new alloc/op   delta
Get/one-of/get-and-make-8                 64.0B ± 0%    128.0B ± 0%  +100.00%  (p=0.008 n=5+5)
Get/one-of/get-only-8                     0.00B          0.00B           ~     (all equal)
Get/less-than-with-int/get-and-make-8     8.00B ± 0%    72.00B ± 0%  +800.00%  (p=0.008 n=5+5)
Get/less-than-with-int/get-only-8         0.00B          0.00B           ~     (all equal)
Get/with-percentage/get-and-make-8         136B ± 0%      200B ± 0%   +47.06%  (p=0.008 n=5+5)
Get/with-percentage/get-only-8            72.0B ± 0%     72.0B ± 0%      ~     (all equal)

name                                   old allocs/op  new allocs/op  delta
Get/one-of/get-and-make-8                  1.00 ± 0%      2.00 ± 0%  +100.00%  (p=0.008 n=5+5)
Get/one-of/get-only-8                      0.00           0.00           ~     (all equal)
Get/less-than-with-int/get-and-make-8      1.00 ± 0%      2.00 ± 0%  +100.00%  (p=0.008 n=5+5)
Get/less-than-with-int/get-only-8          0.00           0.00           ~     (all equal)
Get/with-percentage/get-and-make-8         3.00 ± 0%      4.00 ± 0%   +33.33%  (p=0.008 n=5+5)
Get/with-percentage/get-only-8             2.00 ± 0%      2.00 ± 0%      ~     (all equal)
```